+++ plugins/GSdx/GSDeviceOGL.cpp
@@ -578,7 +578,7 @@ void GSDeviceOGL::DrawIndexedPrimitive()
 
 void GSDeviceOGL::DrawIndexedPrimitive(int offset, int count)
 {
-	ASSERT(offset + count <= m_index.count);
+	ASSERT(offset + count <= (int)m_index.count);
 
 	BeforeDraw();
 	m_state.vb->DrawIndexedPrimitive(offset, count);
